Further separation of tokenisation logic, new method of receiving tokens
