+++ Makefile
@@ -3,7 +3,7 @@ NAME = minishell
 OUT_DIR = output
 
 # Source and Object files
-MINI_SRC = src/parsing/tokenisation_00.c src/parsing/tokenisation_0.c src/parsing/tokenisation_1.c
+MINI_SRC = src/parsing/tokenisation/tokens.c
 MAIN_SRC = src/main.c $(MINI_SRC)
 MAIN_OBJ = $(MAIN_SRC:.c=.o)
 
@@ -13,7 +13,7 @@ TEST_OBJ = $(TEST_SRC:.c=.o)
 
 # Compiler and Flags
 CC = cc
-CFLAGS = -Wall -Wextra -Werror -O3
+CFLAGS = -Wall -Wextra -Werror -g
 TESTFLAGS = -fprofile-arcs -ftest-coverage
 TEST_LIBS = -lcunit
 RM = rm -rf
